HTTP2_PLUS_CLIENT_CERT option is set before openning a connection
