+++ src/libraries/System.Net.Http.WinHttpHandler/tests/FunctionalTests/ServerCertificateTest.cs
@@ -2,31 +2,25 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Net.Security;
-using System.Net.Test.Common;
-using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
-
 using Xunit;
 using Xunit.Abstractions;
 
 namespace System.Net.Http.WinHttpHandlerFunctional.Tests
 {
-    public class ServerCertificateTest
+    public class ServerCertificateTest : BaseCertificateTest
     {
         private readonly ITestOutputHelper _output;
-        private readonly ValidationCallbackHistory _validationCallbackHistory;
 
         public ServerCertificateTest(ITestOutputHelper output)
         {
             _output = output;
-            _validationCallbackHistory = new ValidationCallbackHistory();
         }
+
+        public static bool DowngradeToHTTP1IfClientCertSet => PlatformDetection.WindowsVersion < 2004;
 
         [OuterLoop]
         [Fact]
@@ -34,7 +28,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
         {
             var handler = new WinHttpHandler();
             using (var client = new HttpClient(handler))
-            using (HttpResponseMessage response = await client.GetAsync(System.Net.Test.Common.Configuration.Http.SecureRemoteEchoServer))
+            using (HttpResponseMessage response = await client.GetAsync(Test.Common.Configuration.Http.SecureRemoteEchoServer))
             {
                 Assert.Equal(HttpStatusCode.OK, response.StatusCode);
                 Assert.False(_validationCallbackHistory.WasCalled);
@@ -48,7 +42,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
             var handler = new WinHttpHandler();
             handler.ServerCertificateValidationCallback = CustomServerCertificateValidationCallback;
             using (var client = new HttpClient(handler))
-            using (HttpResponseMessage response = await client.GetAsync(System.Net.Test.Common.Configuration.Http.RemoteEchoServer))
+            using (HttpResponseMessage response = await client.GetAsync(Test.Common.Configuration.Http.RemoteEchoServer))
             {
                 Assert.Equal(HttpStatusCode.OK, response.StatusCode);
                 Assert.False(_validationCallbackHistory.WasCalled);
@@ -67,7 +61,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
                 Assert.Equal(HttpStatusCode.OK, response.StatusCode);
                 Assert.True(_validationCallbackHistory.WasCalled);
 
-                ConfirmValidCertificate(System.Net.Test.Common.Configuration.Http.Host);
+                ConfirmValidCertificate(Test.Common.Configuration.Http.Host);
             }
         }
 
@@ -75,7 +69,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
         [Fact]
         public async Task UseCallback_RedirectandValidCertificate_ExpectedValuesDuringCallback()
         {
-            Uri uri = System.Net.Test.Common.Configuration.Http.RemoteSecureHttp11Server.RedirectUriForDestinationUri(302, System.Net.Test.Common.Configuration.Http.SecureRemoteEchoServer, 1);
+            Uri uri = Test.Common.Configuration.Http.RemoteSecureHttp11Server.RedirectUriForDestinationUri(302, System.Net.Test.Common.Configuration.Http.SecureRemoteEchoServer, 1);
 
             var handler = new WinHttpHandler();
             handler.ServerCertificateValidationCallback = CustomServerCertificateValidationCallback;
@@ -85,7 +79,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
                 Assert.Equal(HttpStatusCode.OK, response.StatusCode);
                 Assert.True(_validationCallbackHistory.WasCalled);
 
-                ConfirmValidCertificate(System.Net.Test.Common.Configuration.Http.Host);
+                ConfirmValidCertificate(Test.Common.Configuration.Http.Host);
             }
         }
 
@@ -99,7 +93,7 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
             handler.ServerCertificateValidationCallback = CustomServerCertificateValidationCallback;
             using (var client = new HttpClient(handler))
             {
-                var request = new HttpRequestMessage(HttpMethod.Get, System.Net.Test.Common.Configuration.Http.SecureRemoteEchoServer);
+                var request = new HttpRequestMessage(HttpMethod.Get, Test.Common.Configuration.Http.SecureRemoteEchoServer);
                 _validationCallbackHistory.ReturnFailure = true;
                 HttpRequestException ex = await Assert.ThrowsAsync<HttpRequestException>(() =>
                     client.GetAsync(System.Net.Test.Common.Configuration.Http.SecureRemoteEchoServer));
@@ -123,100 +117,12 @@ namespace System.Net.Http.WinHttpHandlerFunctional.Tests
             }
         }
 
-        [Fact]
-        public async Task UseClientCertOnHttp2_ClientCertValid_Success()
-        {
-            //Debugger.Launch();
-            await Http2LoopbackServer.CreateClientAndServerAsync(
-                async address =>
-                {
-                    var handler = new WinHttpHandler();
-                    handler.ServerCertificateValidationCallback = CustomServerCertificateValidationCallback;
-                    handler.ClientCertificates.Add(Test.Common.Configuration.Certificates.GetClientCertificate());
-                    handler.ClientCertificateOption = ClientCertificateOption.Manual;
-                    using (var client = new HttpClient(handler))
-                    using (HttpResponseMessage response = await client.SendAsync(new HttpRequestMessage(HttpMethod.Get, address) {Version = HttpVersion.Version20 }))
-                    {
-                        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                        Assert.True(_validationCallbackHistory.WasCalled);
-
-                        //ConfirmValidCertificate(System.Net.Test.Common.Configuration.Http.Host);
-                    }
-                },
-                async s =>
-                {
-                    using (Http2LoopbackConnection connection = await s.EstablishConnectionAsync().ConfigureAwait(false))
-                    {
-                        int streamId = await connection.ReadRequestHeaderAsync();
-
-                        await connection.SendDefaultResponseAsync(streamId);
-                    }
-                }, new Http2Options { ClientCertificateRequired = false });
-        }
-
         private void ConfirmValidCertificate(string expectedHostName)
         {
                 Assert.Equal(SslPolicyErrors.None, _validationCallbackHistory.SslPolicyErrors);
                 Assert.True(_validationCallbackHistory.CertificateChain.Count > 0);
                 _output.WriteLine("Certificate.Subject: {0}", _validationCallbackHistory.CertificateSubject);
                 _output.WriteLine("Expected HostName: {0}", expectedHostName);
-        }
-
-        private bool CustomServerCertificateValidationCallback(
-            HttpRequestMessage sender,
-            X509Certificate2 certificate,
-            X509Chain chain,
-            SslPolicyErrors sslPolicyErrors)
-        {
-            _validationCallbackHistory.WasCalled = true;
-            _validationCallbackHistory.CertificateSubject = certificate.Subject;
-            foreach (var element in chain.ChainElements)
-            {
-                _validationCallbackHistory.CertificateChain.Add(element.Certificate);
-            }
-            _validationCallbackHistory.ChainStatus = chain.ChainStatus;
-            _validationCallbackHistory.SslPolicyErrors = sslPolicyErrors;
-
-            if (_validationCallbackHistory.ThrowException)
-            {
-                throw new CustomException();
-            }
-
-            if (_validationCallbackHistory.ReturnFailure)
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        public class CustomException : Exception
-        {
-            public CustomException()
-            {
-            }
-        }
-
-        public class ValidationCallbackHistory
-        {
-            public bool ThrowException;
-            public bool ReturnFailure;
-            public bool WasCalled;
-            public SslPolicyErrors SslPolicyErrors;
-            public string CertificateSubject;
-            public X509CertificateCollection CertificateChain;
-            public X509ChainStatus[] ChainStatus;
-
-            public ValidationCallbackHistory()
-            {
-                ThrowException = false;
-                ReturnFailure = false;
-                WasCalled = false;
-                SslPolicyErrors = SslPolicyErrors.None;
-                CertificateSubject = null;
-                CertificateChain = new X509CertificateCollection();
-                ChainStatus = null;
-            }
         }
     }
 }
